Adjustments by iurii to #2611 (#2656)
